+++ packages/spacecat-shared-utils/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [@adobe/spacecat-shared-utils-v1.116.2](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.116.1...@adobe/spacecat-shared-utils-v1.116.2) (2026-05-23)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1623](https://github.com/adobe/spacecat-shared/issues/1623)) ([19cb2a4](https://github.com/adobe/spacecat-shared/commit/19cb2a4fec00735ad4eead3030f9ea6b2573fdc3))
+
 ## [@adobe/spacecat-shared-utils-v1.116.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.116.0...@adobe/spacecat-shared-utils-v1.116.1) (2026-05-22)
 
 ### Bug Fixes